requirements.txt
# Data Processing
pandas>=2.0.0
polars>=0.19.0
numpy>=1.24.0
pyarrow>=12.0.0

# Machine Learning
scikit-learn>=1.3.0
lightgbm>=4.0.0

# Datasets
datasets>=2.14.0

# API & Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
slowapi>=0.1.9

# Configuration & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Code Quality
black>=23.11.0
isort>=5.12.0
ruff>=0.1.6

# Logging
structlog>=23.2.0

